shim: Handle shim starting from container or process
We want to be able to start the shim from different parameters.
Indeed, we have a container in case of the create because we want
to update it with the generated PID. But in case of exec, we will
only get a process because there is no need to save this inside
the container.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>