cgroups: Create and delete cgroups files related to provided path
In config.json, a cgroupsPath is provided, and it is expected to be
created. "tasks" and "cgroup.procs" have to be created under this
directory, and filled with the shim's pid. That way, nothing else
will try to create those files on our behalf.

Because we know where the files were created, we are able to remove
them when delete() is called. By removing those files, we make sure
to prevent any file descriptor leak.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>