shim: Configure standards outputs/input before spawning the shim
We want the shim to inherit standards outputs/input from the runtime
but this was not done by default. Indeed, exec.Command() assign by
default the standards out/in to /dev/null.

That's why this patch explicitely assign stdin, stdout and stderr
from the runtime process to the shim process.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>