config: Allow to specify agent pause root path
The agent pause root path was relying on default values to determine
the full pause binary path. This patch allows the user to specify its
own pause root path, through the toml configuration file.

This pause root path is absolute and it is concatenated with the
static relative binary path bin/pause. That way, the configuration
loader can determine a full and usable path.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>