log: Fix the logging to be common to the entire package
The way logrus was used, each file was configured differently. The
behavior we are expecting is to get the same log formatter, log
level, and log output for the whole binary.

That's why we are defining a common logger from the main.go, that
can be used from any file in the same package.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>