vendor: Update virtcontainers vendoring
This is a new update of the virtcontainers vendoring based on latest
changes related to Pod and Process structures. Now the Pod structure
does not hold the proxy URL anymore. The URL is duplicated in every
Process structure because it could be a different one from a
container to another.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>