logging: Log to syslog as well as the logfile.
When run under docker the runtime will be called with
"--log /some/path". However, that path will be container-specific such that if
an error occurs, the log will be deleted making problem diagnosis
impossible.

Enhance the logging so that along with writing to the logfile, the data
also gets written to syslog, which will make it available via
journalctl(1) even if the container setup fails.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>