logging: Add a global log facility.
The global log allows all log output to _also_ be logged to a
non-container-specific path. This is particularly useful under docker
where if a container fails to start, the container-specific directory
(which contains the "--log=" logfile) will be deleted on error, making
debugging a challenge.

The global log comprises one line per entry. Each line contains the
following fields separated by colons (exactly like the "cc-oci-runtime"
global log):

  - timestamp
  - PID
  - program name
  - log level
  - log message

Note that all runtimes share the global log and it is up to the admin to
manage the disk space.

The log can be enabled in the "configuration.toml" config file, or via
the environment variable "CC_RUNTIME_GLOBAL_LOG". The environment
variable takes precedence over the value specified in the configuration
file. The logfile must be specified as an absolute path.

A nice side-effect of this change is that loadConfiguration() only needs
to be called in one place now.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>