oci: Check a cgroup is supported
Because Docker can expect to create several cgroups files for each
cgroup listed by the config.json, we have to make sure this cgroup
is actually supported/mounted.

In case like SemaphoreCI, we cannot control what host kernel is used
and there is no support for "pids" cgroup. That's why this patch makes
sure we are not trying to create new files if the corresponding cgroup
is not mounted.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>