tests: Fix JSON list format tests for go1.9+.
Historically, it has been possible to marshal and demarshal a time.Time and
compare the original and demarshaled values for equality (successfully).

However, go tip (what will presumably become go1.9) changes this behaviour. It
introduces a monotonic element to the time.Time internal representation. Thie
monotonic value (shown as "m=±ddd.nnnnnnnnn" when calling String() on the
time.Time object) is *NOT* retained when marshaling a time.Time.

This ensures comparing the unmarshaled value with the original will fail.

The fix is to call Truncate(0) on the original time.Time object to remove the
monotonic element before attempting comparing with the unmarshaled value.

This change corrects the required "fixup" behaviour in TestStateToJSON() to be
in line with golang expectations.

See: https://go-review.googlesource.com/c/36255/7/src/time/time.go#54

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>