build: Introduce a PREFIX variable
Let's try to keep the level of confusion the lowest possible by adopting
the usual (autoconf) meaning of a prefix. It does mean one has to
specify SYSCONFDIR by hand to preserve the previous behaviour.

It's quite important to do that so make install don't override the user
/etc/clear-containers/configuration.toml which could have been tweaked.

To install system-wide:

$ sudo make PREFIX=/usr SYSCONFDIR=/etc install

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>