run: fix runtime standalone
With this patch a new pts will be created when the console provided as argument
is /dev/ptmx or /dev/pts/ptmx.

Runtime stdout and stdin will be forwarded to the master fd and then to the
slave, something like this:

```
/dev/pts/0 <--> master <--> /dev/pts/1
```

where pts 0 is assigned to the runtime and pts 1 is assigned to the container.

fixes #181

Signed-off-by: Julio Montes <julio.montes@intel.com>