Add CI scripts that SemaphoreCI will use in its project settings
Currently, we hardcoded the commands that SemaphoreCI will
execute for building and testing the project, but any change
requires modification via its web page and we cannot track
any change there.

Therefore, these three scripts will be called in the SemaphoreCI
settings and any change should be done to the scripts instead
of changing the settings.

`setup.sh` will be called in the Setup
`run.sh` will be called on Job 1
`teardown.sh` will be called on After Job

fixes #168

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>