build: Build config file at build time.
Build the config file as part of the default target rather than at
install time. This makes more sense since users may wish to diff(1) the
newly-generated config file with the installed version prior to actually
installing the new one.

Use an order-only prereq for show-summary to ensure "make install"
doesn't rebuild the runtime if it is already current.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>