build: Generate default runtime values.
Generate default values from the build, rather than using hard-coded
values.

This is achieved making the build generate a .go file containing the
default values.

Note that although the conventional way of having the build generate
variable values is to use the linker trick...

  $ go build -ldflags "-X main.foo=bar -X main.a=b ..."

... this is not optimal since:

1) Even though the values are in reality constant, this trick requires
   the go code to define each variable as a 'var' rather than a 'const'
   since constants cannot have their values changed.

2) It doesn't seem to work for "go build" meaning tests fail.

Auto-generating the code before it is even seen by the go tooling
circumvents these issues, plus it is simple and reliable.

Fixes #230.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>