exec: Implement detach option
This patch implements the detach option for the exec command. In case
the flag is set to false, we stay attached to the process. This means
we have to wait for the end of the shim process and return the exit
code through the runtime itself.

Docker already implements this option on his own, meaning this patch
is useful for cases where the runtime is used as a standalone binary.

For instance, CRI-O uses the runtime with this option, meaning it
expects this to work as expected.

Fixes #318

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>