build: Add standard targets.
The default prefix is /usr/local. This is conventional but does generate
the required configuration for most developers (who are building the
runtime on systems that use OBS-generated packages).

Introduce two new targets that will "do the right thing" for the
developers:

  - build-standard
  - install-standard

Note that the "standard build" can be triggered by setting the
environment variable "CC_STANDARD_BUILD" to any value. For example,
rather than using the new targets, you could run:

  $ export CC_STANDARD_BUILD=yes
  $ make && sudo -E sh -c 'make install'

Note further that if you wish to see how a standard build will work,

  $ make help CC_STANDARD_BUILD=yes

Finally, not that summary generated by the build will display if the build is
a standard one.

Fixes #275.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>