run: Implement console-socket support
This patch implements the new --console-socket option for our runtime
and it also cleans some unneeded code.
Particularly, setting a console in case there is no console path or
console socket provided, is not needed. Indeed, in that case, we have
the shim inheriting stdin, stdout and stderr. The RAW terminal does
not have to be set from the runtime side, it has to be set inside the
agent.

Fixes #317

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>