console: Don't link the shell to the shim if detach and tty are enabled
In case our runtime is used in standalone mode with the --detach
and tty flags enabled, we should not get any output on the current
shell. We should still be able to get some outputs if a console has
been provided though.

This patch relies on the latest virtcontainers vendoring to be
able to indicate if the process is expected to be detached.

Fixes #326

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>