tests: Add framework to mock virtcontainers
Add a simple test framework to allow the behaviour of the virtcontainers
package to be mocked.

This change introduces a RVC interface and provides two implementations:
one for virtcontainers itself (rvc-implementation.go) and another mock
implementation for the tests to allow the behaviour of the
virtcontainers package to be modified (rvc-implementation_test.go). By
only accessing the virtcontainers implementation via the "vci" variable,
the tests can be switched to the mock implementation.

Note that this mocking is used rather than testing virtcontainers
directly since the virtcontainers API is already tested in that package:
the mock framework thus avoids duplication of test code and allows more
fine-grained error scenarios to be handled.

This change also introduces a single new test, TestCreate(), that
increases the unit test coverage of create.go by using the new framework
to force create() to fail due to a virtcontainers failure in
CreatePod().

Fixes #401.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>