delete: A pod has to be stopped from delete command
In case of a container, if kill has been previously called and has
terminated the container's process, we now have virtcontainers
updating the container state. But the pod state will never switch
to "stopped" implicitly. We have to stop the pod on purpose to end
properly the lifecycle. This means in case delete command is called
for the pod itself (Docker or CRI-O pod cases), we have to call into
StopPod() no matter what is the value of the force flag.

Fixes #377

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>