cli: Check for empty string for cli flag "process" in exec.go
Instead of checking for the "process" flag value with IsSet,
check if the flag's value is an empty string instead. The IsSet
call returns true only if the flag is passed explicitly on the
command line.
Hence values set by other means such as default values,
environment etc will be ignored completely.

This will also help with testing, as the "process" flag can now
be set in the unit tests.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>