tests: Run unit tests as root and non-root.
Some of the unit tests require root privileges to run. Conversely, some must
be run as a non-root user.

Modify the unit-tests to run twice, once as a non-root user and again
as the root user, and combine the resulting unit-test coverage files to
get a true unit-test coverage figure.

Fixes #483.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>