config: Support a stateless system
Modify the runtime to look in multiple locations for its configuration.

Previously, with a CC system build, the build system would generate (and
the runtime would read) the file below:

    /etc/clear-containers/configuration.toml

This file was represented by the DESTCONFIG build variable.

However, now a CC system build will generate the following path instead:

    /usr/share/defaults/clear-containers/configuration.toml

But the runtime will look in both locations in the following order
(with build variables in brackets):

    /etc/clear-containers/configuration.toml (DESTSYSCONFIG)
    /usr/share/defaults/clear-containers/configuration.toml (DESTCONFIG)

(Note that the original configuration file now has a new build variable
name: DESTSYSCONFIG).

The first file the runtime finds will be used. Since previously the
configuration file below "/etc" was used and now the build will generate
the file below "/usr/share/defaults" instead, the advice is to delete
the "/etc/clear-containers" directory to ensure that the runtime uses
the latest pristine built configuration file by default.

Since the runtime requires a configuration file to run most commands,
but since the runtime is now looking in multiple locations for this
configuration, this change also introduces a new
"--cc-show-default-config-paths" global option. This option simply
displays the list of paths the runtime will consider as a configuration
source and then exits. This option does not require a configuration file
and may be useful to establish where the runtime is obtaining its
configuration from.

Fixes #551.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>