Debug: Enable debug logs from Hypervisor and Container Kernel
Enable hypervisor and container kernel logs. Once debug is
enabled these logs can be obtained in the cc-proxy logs or
by stopping the cc-proxy service and running the cc-proxy
explicitly.

/usr/libexec/clear-containers/cc-proxy -log debug

This option is useful to trouble shoot errors in the kernel
or when the agent in the container fails to launch.

Fixes https://github.com/clearcontainers/runtime/issues/568
Fixes #568

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>