Configuration: Provide an option to disable nesting customizations
Provide an option to disable nesting customization. This is useful
when running on top of KVM, where disabling the customizations
performed when the runtime detects it is running on top of a VM
results in higher performance.

This will also be an useful switch to disable the nesting
workarounds once the KVM issues are addressed without requiring
any code changes on any existing deployments.

Fixes #603.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>