Configuration: Huge pages support
Enable huge pages support at a node level. This results in all
containers on that node being launched using huge pages.

This is useful when these nodes are used for NFV workloads
where there will be user mode networking stacks in use on that
host or within the container.

Fixes #626

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>