list: Ensure "--cc-all" details are correct
If a container was using old assets (kernel + image) and those assets
were then updated, "list --cc-all" would show, incorrectly, that the
existing container was using the latest asset versions, not the correct
older versions.

Removed a couple of now-incorrect tests and added a new one to assert
the correct list behaviour.

Fixes #628.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>