scripts: Add script to find stale containers
A stale container is one which was running at the time assets (kernel +
rootfs image) were updated: they are using older asset versions than
are currently available on the system.

Provide a script to determine these stale containers to allow an
administrator to decide if they should be restarted to take advantage
of the new asset versions.

Fixes #630.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>