versions: Add missing components
This commit adds the components (and their respective version) that
are listed in the clearcontainers/tests/test-versions.txt. After
merging this, we would be able to remove the test-versions.txt
from the tests repository and rely in the versions.txt from this
repository, which will make maintenance easier and less
fragile.

The components added this list are:
- Qemu-lite
- CRI-O
- runc for CRI-O
- kubernetes
- Openshift
- Go version

Fixes: #608.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>