exec: Fix "docker exec --user"
Fix running `docker exec` with the `--user=` option. Previously,
specifying `--user=` with any user (or user and group) would be ignored
and the workload would run as `root` (`UID`+`GID`). This was because the
runtime was setting the user to run the workload from the `Username`
field in the OCI `Process.User` struct. However, that field is
Windows-only so the value was always zero, denoting `root`.

Fixes #867.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>