config: Add the agent field by default
Instead of relying on the implicit default agent being cc for Clear
Containers, this commit adds the agent field to configuration.toml
file so that any user will be able to make sure Clear Containers is
the agent used by default.

Also, this will allow for changing easily the agent type from cc to
a new one like kata for Kata Containers.

Fixes #903

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>