build: Allow for kata project build
Rework the build system to allow the runtime to be built for either
of the following:

- Clear Containers (`cc-runtime` binary and configuration)
- Kata Containers (`kata-runtime` binary and configuration)

The existing build rules are retained but there are now two new rules:

```
$ make build-kata-system
$ sudo make install-kata-system
```

Two new variables have also been added:

- `KATA_SYSTEM_BUILD`: set to any value to build/install for Kata.
- `SYSTEM_BUILD_TYPE`: build for either Clear Containers (set to `cc`) or
  Kata (set to `kata`).

See `make help` for further details.

Note that configuration files are installed to a project-specific
directory, allowing for systems to be installed for both Kata and Clear
Containers.

Finally, the bash completion and collection scripts have been updated
to work for either project.

Fixes #919, #921.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>