kill: Fix CRI-O race condition
CRI-O assumes that anything related to the container should be able
to be unmounted after the container process returned after a kill.
In our case, we have a strong dependency on what should be done before
CRI-O can actually unmount storage because it is plugged into the VM.

This means we have to translate a kill as a stop when needed. The
condition for that is to detect when someone calling into kill would
actually expect to see the container completely stopped. This case
happens for SIGKILL and SIGTERM, and those are the conditions for
our runtime to consider stopping the container after it has been
signalled.

Fixes #1002

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>