logging: Add CLI command to logs
Since the runtime is called multiple times by the container manager, add
the command-line sub-command name to all log entries to make it clear
which command was running when a particular log message was produced.

Note that CLI arguments are already logged once at runtime startup. This
change is more useful for the scenarios where only a log excerpt is
available (for example when running `cc-collect-data.sh` which only
extracts error logs).

Fixes #1024.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>